tests/wpt/web-platform-tests/xhr/open-url-multi-window-2.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: open() resolving URLs (multi-Window; 2; evil)</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="https://xhr.spec.whatwg.org/#the-open()-method" data-tested-assertations="following::ol[1]/li[2]/ol[1]/li[1]" />
  </head>
  <body>
    <div id="log"></div>
    <script>
    function init(){ // called from page inside IFRAME
        test(function() {
          var client = new self[0].XMLHttpRequest();
          let exceptionCtor = self[0].DOMException;
          document.body.removeChild(document.getElementsByTagName("iframe")[0])
          assert_throws_dom("InvalidStateError", exceptionCtor, function() {
            client.open("GET", "folder.txt")
          }, "open() when associated document's IFRAME is removed");
        });
      }
    </script>
    <iframe src="resources/init.htm"></iframe>
 </body>
</html>
